Knausj generic terminal establishment (#514)
* changed kde_konsole app name to lower case konsole

* moved kde-konsole integration to right folder

* added todos to konsole integration

* added tab previous as alternative to tab last

* marked platform specific terminal.talon files as deprecated/WIP

* replaced generic_terminal tag by terminal tag

* added kde-konsole implementation

* removed unnecessary imports from kde-konsole implementation

* added rerun search to generic_terminal.talon

* added generic_terminal to generic grammars in README

* updated TODOs in generic_terminal.talon

* removed apple_terminal-> fully merged into generic_terminal

* readded terminal.talon for apple (fitted to generic_terminal)

* removed "resume" todo from generic_terminal

* removed deprecated generic_terminal tag from powershell

* added implementation for rerun search to powershell

* removed unused variable from power_shell.py

* removed unnecessary import from power_shell.py

* moved power_shell tags to talon file for consistency

* fixed word_left and right for linux terminal.py

* added explanation for cancel and ignore

* added anaconda and kubectl tag to kde konsole

* added generic unix and windows shell files

* fixed power_shell.talon

* fixed errors with generic_windows_shell tag

* improved explanation for terminal tags

* renamed linux_terminal.py and moved tabs implementation

* removed outdated comment

* removed redundant implementations

* fixed find in linux_terminal.py

* fixed tab jumping for firefox on linux

* removed linux_terminal.py and moved contents to specific terminals

* added TODOs to kde konsole

* improved kde-konsole implementation

* added generic_<>_shell tags to tags.py

* fixed activation of generic_<>_shell files

* removed unused code from kde-konsole.py
