+++ apps/generic_terminal/generic_terminal.talon
@@ -1,5 +1,7 @@
-tag: user.generic_terminal
+tag: terminal
 -
+# tags should be activated for each specific terminal in the respective talon file
+
 lisa: 
     user.terminal_list_directories()
 lisa all: 
@@ -8,4 +10,6 @@ katie [<user.text>]: user.terminal_change_directory(text or "")
 katie root: user.terminal_change_directory_root()
 clear screen: user.terminal_clear_screen()
 run last: user.terminal_run_last()
+rerun [<user.text>]: user.terminal_rerun_search(text or "")
+rerun search: user.terminal_rerun_search("")
 kill all: user.terminal_kill_all()
