--- apps/linux/kde-konsole.talon
@@ -1,5 +1,0 @@
-os: linux
-app.name: Konsole
--
-tag(): terminal
-tag(): user.git
