+++ apps/platforms/linux/kde-konsole/kde-konsole.talon
@@ -1,0 +1,18 @@
+os: linux
+and app.name: konsole
+-
+# makes the commands in generic_terminal available
+tag(): terminal 
+
+# activates the implementation of the commands/functions in generic_terminal
+tag(): user.generic_unix_shell
+
+# makes commands for certain applications available
+# you can deactivate them if you do not use the application
+tag(): user.git
+tag(): user.anaconda
+# tag(): user.kubectl
+
+
+tag(): user.tabs
+# TODO: add file_manager support
