+++ apps/platforms/mac/terminal/terminal.py
@@ -1,6 +1,8 @@
 from talon import Context, Module, actions, imgui, settings, ui
 import os
 
+# TODO: fit this to generic_terminal
+
 ctx = Context()
 ctx.matches = r"""
 app: apple_terminal
