+++ apps/platforms/win/powershell/power_shell.talon
@@ -1,0 +1,16 @@
+os: windows
+and app.name: Windows PowerShell
+os: windows
+and app.exe: powershell.exe
+-
+# makes the commands in generic_terminal available
+tag(): terminal 
+
+# activates the implementation of the commands/functions in generic_terminal
+tag(): generic_windows_shell
+
+# makes commands for certain applications available
+# you can deactivate them if you do not use the application
+tag(): user.git
+tag(): user.anaconda
+# tag(): user.kubectl
